requirements.txt
boto3==1.17.49
colorlog==5.0.1
click==7.1.2  # Required because version 8 is incompatible with typer
dvc[s3]==2.5.0
jupyter==1.0.0
kaleido==0.2.1  # Needed for optuna
lightning-bolts==0.3.4
matplotlib==3.4.2
numpy==1.20.3
optuna==2.8.0
pandas==1.2.4
Pillow==8.2.0
plotly==4.14.3  # Needed for optuna
pyarrow==4.0.1
pytest==6.2.4
pytorch-lightning==1.3.8
ruamel.yaml==0.17.10
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
sunpy[net]==2.1.5
torch>=1.8.0+cu111
torchmetrics==0.4.0
torchvision>=0.9.0+cu111
torch_optimizer==0.1.0
tqdm==4.61.1
git+https://github.com/jacobgil/pytorch-grad-cam

-e .

--find-links https://download.pytorch.org/whl/torch_stable.html
